WIP attempt at implementing BotCommand in a cool way
I thought BotCommand could directly contain references into the framework's command objects... but this turned out to be a MAJOR FUCKING PAIN because then the framework instance has to be passed into every goddamn place where arguments are passed, and this brings a bunch of stupid problems with it. Just looks at slash/argument.rs, it's atrocious. The coolness of this approach isn't worth it